feat: Add CRSF protocol support in control logic
This commit adds support for the CRSF protocol in the control
logic of the "rcfw.ino" file. The code now conditionally includes
either the IBusBM or AlfredoCRSF library based on the SERIAL_PROTOCOL
define.

Changes include:
- Added new `#define` for CRSF and conditional logic to switch
  between using IBusBM and AlfredoCRSF.
- Adjusted setup and control functions to support CRSF.
- Added debugging prints for CRSF to monitor channel data.
- Updated logic in `loop()` to conditionally handle and read input
  from either the IBus or CRSF based on the protocol selected.

These updates allow the firmware to be used with multiple protocols,
increasing the functionality and flexibility of the software.