refactor: Remove unused failsafe code
Remove the unused failsafe_channel declaration and related
conditional failsafe code in the loop function. These elements
were not being utilized in the current implementation for either
IBus or CRSF protocols, making this code redundant.

Originally this was part of an attempt at implementing a failsafe for
receivers not properly supporting it - but never quite worked. ELRS
receivers should always be able to do proper failsafe handling on
receiver side, so use one of those if you need it.